pyproject.toml
[project]
name = "tiet-notegen-v2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bs4>=0.0.2",
    "demjson3>=3.0.6",
    "dotenv>=0.9.9",
    "google-genai>=1.24.0",
    "pydantic>=2.11.7",
    "pymupdf>=1.26.1",
    "requests>=2.32.4",
    "rich>=14.0.0",
    "textual>=3.5.0",
]
